Added DeepSlice Library
This added the DeepSlice Library to the project.
There are still some changes needed to be made to be fully integrated, but it can now successfully call the DeepSlice libraries and save the results as .csv and .xml files.
For now, it should be a good base for implementing further features.